Persist activities between page transitions
The main update in this commit was changing the activities attribute to be a Set rather than a List. To accomplish this, the instances in which parts of the app needed to read in a list of activities, like the activities.dart and single_day_summary.dart pages, there needed to be a getter method to return a list of the Day's Set of activities. The new_summary.dart page created several Day objects, so each instantiation needed to input a Set rather than a List of activities to the Day constructor. Finally, the activities.dart page needed additional logic to update the Day object and keep the 'answers' HashMap updated with the same values in the Day's Set of activities. Since the 'answers' HashMap mirrored the Set of activities, the checkboxes on the page now persist and remain up-to-date.