Resolve race condition involving reading new Days from the database immediately after writing a new one
Change the addDay method in backend_service.dart to return a File object. Modify questions_controller.dart to only navigate to the new_summary.dart page when the File is returned from the addDay method. Delete the temporary solution in new_summary.dart that previously added the missing Day if it did not exist in the list of Days returned from the database.