Save the user's name with SharedPreferences
Change the addName function to updateName since the function will also replace an existing name if one already exists. Also update the getName and updateName functions in the backend service to save and retrieve the user's name with SharedPreferences. For the two previous implementations of the updateName method (on the signup_screen and user_settings pages), the application waited for the name to be retrieved back from the backend. Instead, this function now returns 'void' and can be executed asynchronously.